Make sure task countdown is an int.
I have a suspicison that having it as a float is causing it to be treated as 0.

+++ pingapp/views.py
@@ -35,7 +35,7 @@ def spawn_pings(request, frequency):
     #evenly over the hour/day.
     if not urls_to_ping:
         return HttpResponse('No URLs to ping')
-    interval = math.floor(float(TIME_PERIODS[frequency]) / float(len(urls_to_ping)))
+    interval = int(math.floor(float(TIME_PERIODS[frequency]) / float(len(urls_to_ping))))
     delay = 0
     for url in urls_to_ping:
         querystring = urlencode([('url', url)])
